C type `long` is used for array dimensions, number of elements and indices
I decided to use the same rule of thumb as in Yorick as (i) `long` integer
type is usually 32 bits on 32-bit machines and 64 bits machines and
(ii) `long` is a signed integer type which is very useful (compared to, e.g.,
`size_t`) to detect errors and to compute offsets.

+++ yorick/yor-tao-rt.c
@@ -808,7 +808,7 @@ Y_tao_set_data(int argc)
         y_error("not same number of dimensions");
     }
     for (d = 1; d <= ndims; ++d) {
-        size_t dim = tao_get_shared_array_size(arr, d);
+        long dim = tao_get_shared_array_size(arr, d);
         if (dims[d] != dim) {
             y_error("not same number of dimensions");
         }
